### 2018-08-04 #840 @arfon
👋@dleehr - could you take a look at this list of reviewers and identify a few potential candidates? https://bit.ly/joss-reviewers
