### 2018-08-07 #840 @arfon
> Sure - I don't see a lot of overlap between the technologies (Ruby on Rails) and the domain (evolution/trait data) in the reviewers. Considering those and the focus on dataset homogenization, I'd suggest @HadrienG, @dredamonsta1, @amoeba, @krother. Thanks!

Yes, this might be tough :smile:. I know a fair amount about Ruby on Rails and so might be able to help with some of the review aspects there (we may want to assign a different editor in that case).

@HadrienG, @dredamonsta1, @amoeba, @krother - would you be willing to review this submission for JOSS?
