### 2018-08-21 #840 @pjotrp
Hi @dleehr, even better I can be the editor ;). Before I look for reviewers I would like you to work a little on the paper. It is too short now. If you check other JOSS papers you can see we like short, but it would be good to add more context, prior art, and future work. Look at one of mine, for example, http://joss.theoj.org/papers/10.21105/joss.00025.

The idea is to put your project in the best possible light so people start using it. Right? People like figures, so a screen shot will help.
