### 2018-12-28 #840 @pjotrp
@dleehr it should not be too hard. Just make clear who uses the project today and what scientific context it has. Alternatively pull the submission. It has to be worth the effort, right? Also for the reader (take pity on the reader!)
