### 2019-01-22 948d8fb
Rewrite Summary in paper.md
