### 2019-01-25 #840 @pjotrp
@amoeba would you be able to review in the coming weeks?
