### 2019-02-01 03c37ca
Update .travis.yml to use ruby 2.2.0
