### 2019-02-08 NESCent/TraitDB#222 @amoeba
JOSS Review
Hi @dleehr, I've had a chance to look over TraitDB and I appreciate the level of polish you've put on the documentation and the software as a whole. As the reviewer guidelines direct, I'm leaving comments specific to my review of your JOSS submission https://github.com/openjournals/joss-reviews/issues/1201 here.

I got a bit stuck at installation (configuring auth, really). Perhaps you could offer some assistance here.

When I try to log in with Google, I get:

I edited `devise.rb` and manually set the client ID and secret and login worked after that (I also had to set the callback URL in the Google Console). I'm running `rails server` in a shell with the two environmental variables set. I set my PostgreSQL environmental variables the same way and they were picked up fine. I'm using rbenv to manage rubies. What do you use?

**Other notes:**

- Your `Gemfile` specifies Ruby 2.2.0 but Bundler requires Ruby 2.3.0. I got an error from Bundler about this when running `bundle install` and worked around it by removing the Ruby version restriction from the `Gemfile`. Upgrading from Ruby 2.2.0 to 2.3.0 is quite a change for a JOSS review to request but, since others might encounter this, it might be a good idea to do at some point in the future.
- Get an OpenID link on login page links to "https://openid.net/get-an-openid/" which is a 404
- Without logging in, there is no Upload button. I think the text in the readme should be updated to reflect that.

- After logging in and upgrading my user to an admin, creating a new project worked. I then went to Upload, clicked Import Configs, clicked Upload New CSV Import Config, clicked Choose File, selected `lib/traitdb_import/Sample.yml`, clicked Upload CSV Import Config and got an error:

    >1 error prohibited this record from being saved
    >There were problems with the following fields:
    >
    >Config file has contents that are not what they are reported to be

Any thoughts on my two issues here? I can take a look again tomorrow but have to run for now.
