### 2019-02-08 #1201 @amoeba
Hi @pjotrp I started my review but encountered some issues I didn't expect (documented on https://github.com/NESCent/TraitDB/issues/222) so I'll need some more time for my review.
