### 2019-02-13 NESCent/TraitDB#222 @dleehr
Thanks for the feedback and for digging pretty deep to get past some holes in the documentation. I'll try to address those, but will be away next week.

I too use rbenv and will try upgrading to 2.3.0. Will certainly address the issues. Thanks @amoeba !
