### 2019-02-18 NESCent/TraitDB#222 @amoeba
Regarding my issues with Devise earlier, I figured out that I was setting my env vars wrong. I use fish shel and I needed to add a `-x` to my `set` in order for the env vars to be shared with subprocesses. Installation work as documented.

Regarding my config upload issue, I tried again to get a config to upload and no matter which one I select I get:

> Config file has contents that are not what they are reported to be

I've tried the configs in `lib/traitdb_import` and also the example on the wiki with no luck. Am I doing something wrong here?

I ran through the other checklist items and have a couple of things. Here's the complete list of outstanding review items:

- **Version: Does the release version given match the GitHub release (v0.9.1)?**

  No, but this can be resolved once everything else is done.

- **Functionality: Have the functional claims of the software been confirmed?**

  Not yet, as I haven't gotten fully running due to not being able to import a config.

- **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

    Could you please add this to the readme?

* **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

  - `ashman2014tree ` needs its DOI, https://doi.org/10.1038/sdata.2014.15

Everything else looks good. Could you please take a look at my issue with configs and let me know if I should be doing something differently? Thanks!
