### 2019-02-26 NESCent/TraitDB#222 @dleehr
Yep, will look at the configs today @amoeba. Thanks for your patience and diligence on the installation issues!, I was on a family vacation last week,
