### 2019-03-05 b65e6f0
Update README with database user creation step
