### 2019-03-05 NESCent/TraitDB#222 @dleehr
Hey @amoeba I fixed the config loading issue in #225 and verified uploading `SampleDataset.csv` with `Sample.yml` config. However I discovered that the result display is broken (#226) so am working on that now.
