### 2019-03-05 485362d
Clarify README upload section to mention logged-in requirement
