+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+## How to contribute to TraitDB
+
+#### **Did you find a bug?**
+
+* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/NESCent/TraitDB/issues).
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/NESCent/TraitDB/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible demonstrating the expected behavior that is not occurring.
+
+#### **Did you write a patch that fixes a bug?**
+
+* Open a new GitHub pull request with the patch.
+
+* Ensure the PR description clearly describes the problem and solution. Include the relevant issue number if applicable.
+
+* Ensure the patch includes unit tests and does not break existing tests.
+
+Thanks!
