### 2019-03-05 NESCent/TraitDB#222 @dleehr
I've fixed the issues here, and released v1.0.0. I believe this is ready for another look now @amoeba
