### 2019-03-06 NESCent/TraitDB#222 @amoeba
Sounds great @dleehr. I can take another look before next week.
