### 2019-03-11 NESCent/TraitDB#222 @amoeba
Hey @dleehr, everything looks great now. I was able re-run initial setup, upload a config and matching data, and test out search. I've scanned the edits you've made to address my remaining issues and everything's satisfactory. I've updated the checklist over on https://github.com/openjournals/joss-reviews/issues/1201 and am about to walk over and ask the editor to finalize the submission.

Thanks for taking the time to work through my edits!
