### 2019-03-11 #1201 @amoeba
Hi @pjotrp, @dleehr has addresses all of my outstanding issues with the review over on https://github.com/NESCent/TraitDB/issues/222 and I've updated the checklist above to reflect that. Everything is in order for your final stamp except

> Version: Does the release version given match the GitHub release (v0.9.1)?

which should be updated to 1.0.0 (See: https://github.com/NESCent/TraitDB/releases/tag/v1.0.0).

My review is: Accept (conditional upon the above)
