### 2019-03-11 #1201 @pjotrp
Thank you @amoeba! @dleehr to finalize your submission and accept your paper in JOSS, we need two things. First, can you confirm that all references in your bibliography have a DOI in the bibliography (if one exists).

Second, we need you to deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-archiving service.

To do so:

3. Post a comment here with the DOI for the release.
