### 2019-03-11 #1201 @dleehr
Thanks @pjotrp

> First, can you confirm that all references in your bibliography have a DOI in the bibliography

Yes

> Second, we need you to deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-archiving service.

https://doi.org/10.5281/zenodo.2583988
