### 2019-03-12 #1201 @labarba
Hi @dleehr — Please edit the metadata of the Zenodo deposit (note: no need for a new DOI or version) so the title and author list match the paper.

The paper proof has a References section, but no references listed. Can you check that?

Editorial suggestions/fixes:
- you could merge your first four mini-paragraphs into one paragraph.
- "While these software are technically proficient can facilitate robust operations…" >> something is wrong with this phrase
- "TraitDB aids the ingest of datasets">> grammar? if "ingest" is a verb, why the article?

To be honest, after reading the paper, it's hard to visualize wha the software does. I remind you that the JOSS paper should be understandable to non-specialists. Can you try to improve the readability, overall, and state more clearly what the software _does_?
