### 2019-03-12 #1201 @pjotrp
@labarba some valid comment. Except for the last one. I checked the paper for having a topical introduction for laymen. It says that it stores data from different formats using templates.  What else do you want to add?
