### 2019-03-12 #1201 @labarba
Since the author is requested to make some edits, seems like a final effort to make it read easier is not too much to ask.
