### 2019-03-25 ab2b274
Rewrite intro paragraph
