### 2019-03-26 544cd12
Rewrite opening paragraph to be more cohesive
