### 2019-03-26 ca085d3
Remove empty References heading
