### 2019-03-26 #1201 @dleehr
Hey @pjotrp, I revised the intro paragraph to move some information up and state the case more clearly to a general audience.

I registered a new DOI at Zenodo: 10.5281/zenodo.2609163, since the auto-generated github submission did not allow any edits (e.g. to correct the author list)
