### 2019-03-27 #1201 @labarba
The XML deposit includes one citation, whereas the paper has no references.

@arfon : Why does `whedon` find a citation when there are none? Is it picking it up from a `.bib` file? In that case, should the `.bib` file be deleted? Is the citation included in the XML going to cause a false citation entry with Crossref?
