### 2019-03-28 #1201 @arfon
> @arfon : Why does whedon find a citation when there are none? Is it picking it up from a .bib file? In that case, should the .bib file be deleted? Is the citation included in the XML going to cause a false citation entry with Crossref?

Yes please remove the entry from the bibtex file (but don't delete it).
